# docs/swagger.yaml
basePath: /api/v1
definitions:
  types.CreateUserRequest:
    properties:
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      user_name:
        type: string
    type: object
  types.EditUserRequest:
    properties:
      dob:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      user_id:
        type: integer
    type: object
  types.LoginRequest:
    properties:
      password:
        type: string
      user_name:
        type: string
    type: object
  types.MessageResponse:
    properties:
      message:
        type: string
    type: object
  types.PostDetailResponse:
    properties:
      content_image_path:
        type: string
      content_text:
        type: string
      created_time:
        type: string
      post_id:
        type: integer
      user_id:
        type: integer
      visible:
        type: boolean
    type: object
info:
  contact:
    email: tpdongcs@gmail.com
    name: Dong Truong
    url: https://www.linkedin.com/in/dong-truong-56297a145/
  description: A simple social network management service API in Go using Gin framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Social network Service
  version: "1.0"
paths:
  /posts:
    post:
      consumes:
      - application/json
      description: create new post
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: create new post
      tags:
      - Post
  /posts/{post_id}:
    delete:
      consumes:
      - application/json
      description: delete post by post id
      parameters:
      - description: post id
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: delete post
      tags:
      - Post
    get:
      consumes:
      - application/json
      description: get post detail by post id
      parameters:
      - description: post id
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PostDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: get post detail
      tags:
      - Post
    put:
      consumes:
      - application/json
      description: edit post by post id and new content text or new content image
        path or new visible
      parameters:
      - description: post id
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: edit post
      tags:
      - Post
  /posts/{post_id}/comments:
    post:
      consumes:
      - application/json
      description: create post comment by post id and user id and content text
      parameters:
      - description: post id
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: create post comment
      tags:
      - Post
  /posts/{post_id}/likes:
    post:
      consumes:
      - application/json
      description: like post by post id
      parameters:
      - description: post id
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: like post
      tags:
      - Post
  /users:
    post:
      consumes:
      - application/json
      description: create new user using user provided information
      parameters:
      - description: create user param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: create user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: edit user using user provided information
      parameters:
      - description: edit user param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.EditUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: edit user
      tags:
      - User
  /users/login:
    post:
      consumes:
      - application/json
      description: check user user_name and password
      parameters:
      - description: login param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: check user authentication
      tags:
      - User
securityDefinitions:
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: Grants read and write access to administrative information
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Password:
    flow: password
    scopes:
      admin: Grants read and write access to administrative information
      read: Grants read access
      write: Grants write access
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"